+++ src/app/productListComponent/productListComponent.component.ts
@@ -1,5 +1,7 @@
 import { Component, Input, Output, EventEmitter } from '@angular/core';
 import { Product } from '../../core/productListComponent/products.interface';
+import { ProductsService } from 'src/core/productListComponent/productService';
+import { CartService } from '../../core/productListComponent/cartService';
 @Component({
   selector: 'shop-productListComponent',
   templateUrl: './productListComponent.component.html',
@@ -7,6 +9,15 @@ import { Product } from '../../core/productListComponent/products.interface';
 })
 
 export class ProductListComponent {
-  @Input() products!: Product[];
-  @Output() buttonClick = new EventEmitter<Product>();
+  products!: Product[];
+
+  constructor(private productService: ProductsService, private cartService: CartService) { }
+
+  ngOnInit(){
+    this.products = this.productService.getProducts();
+  }
+
+  addToCart(product: Product): void {
+    this.cartService.addToCart(product)
+  }
 }