+++ src/core/productListComponent/productServices.ts
@@ -1,7 +1,0 @@
-import { productsListMock } from './productServices.mock';
-
-export class ProductsService {
-  getProducts(){
-    return productsListMock
-  }
-}